CXP-1100: Move controllers
