+++ src/Akeneo/Connectivity/Connection/back/tests/PhpStan/Rule/SuffixControllerByActionRule.php
@@ -10,7 +10,7 @@ use PHPStan\Analyser\Scope;
 
 final class SuffixControllerByActionRule extends AbstractControllerRule
 {
-    private const ERROR_MESSAGE = 'Controller must be suffixed with "Action"';
+    private const ERROR_MESSAGE = 'ExternalApi must be suffixed with "Action"';
 
     /**
      * @param Class_ $node
