+++ src/Akeneo/Connectivity/Connection/back/tests/Unit/spec/Infrastructure/Apps/Controller/ExternalApi/RequestAccessTokenActionSpec.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace spec\Akeneo\Connectivity\Connection\Infrastructure\Apps\Controller;
+namespace spec\Akeneo\Connectivity\Connection\Infrastructure\Apps\Controller\ExternalApi;
 
 use Akeneo\Connectivity\Connection\Application\Apps\Service\CreateAccessTokenInterface;
 use Akeneo\Platform\Bundle\FeatureFlagBundle\FeatureFlag;
